Move new integration tests to resolve import cycle
