+++ integrationtests/integration_test_net.go
@@ -1,4 +1,4 @@
-package tests
+package integrationtests
 
 import (
 	"context"
